chore: move legacy app loader and fix loading
