+++ client/apps/fusion-app-loader/app.config.ts
@@ -1,0 +1,24 @@
+import { mergeAppConfigs, defineAppConfig } from '@equinor/fusion-framework-cli';
+export default defineAppConfig((_env, { base }) =>
+	mergeAppConfigs(
+		{},
+		{
+			environment: {
+				endpoints: {
+					client: {
+						baseUri: 'https://backend-fusion-project-portal-test.radix.equinor.com',
+						defaultScopes: ['api://02f3484c-cad0-4d1d-853d-3a9e604b38f3/access_as_user'],
+					},
+					portal: {
+						baseUri: 'https://fusion-s-portal-ci.azurewebsites.net',
+						defaultScopes: ['5a842df8-3238-415d-b168-9f16a6a6031b/.default'],
+					},
+				},
+				fusionEnv: 'ci',
+				loadingText: 'Loading',
+				appKey: 'handover-garden',
+				basename: 'localhost:3000/apps/handover-garden/',
+			},
+		}
+	)
+);
