+++ client/apps/fusion-app-loader/app.manifest.config.ts
@@ -1,0 +1,7 @@
+import { defineAppManifest, mergeManifests } from '@equinor/fusion-framework-cli';
+
+export default defineAppManifest((env, { base }) => {
+	return mergeManifests(base, {
+		name: 'App-loader',
+	});
+});
