+++ client/apps/fusion-app-loader/package.json
@@ -1,0 +1,48 @@
+{
+  "name": "fusion-app-loader",
+  "version": "1.0.0",
+  "description": "",
+  "type": "module",
+  "main": "src/index.ts",
+  "scripts": {
+    "prebuild": "rm -f ../../../client/packages/portal-client/src/assets/appLegacyLoader.js",
+    "build": "fusion-framework-cli app build && mv ./dist/app-bundle.js ../../../client/packages/portal-client/src/assets/appLegacyLoader.js",
+    "build:resources": "tsc && vite build",
+    "dev:appLoader": "fusion-framework-cli app dev -c vite.config.js",
+    "docker": "cd .. && sh docker-script.sh app-react",
+    "update:fusion": "ncu -i -f /fusion/"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "@equinor/eds-core-react": "^0.42.0",
+    "@equinor/fusion": "^3.4.16",
+    "@equinor/fusion-components": "^2.12.0",
+    "@equinor/fusion-framework-app": "^9.1.10-next-9f6eec8c5a4cb3dd430fc6bee227442a3dc28bb8",
+    "@equinor/fusion-framework-module-app": "^6.0.0-next-9f6eec8c5a4cb3dd430fc6bee227442a3dc28bb8",
+    "@equinor/fusion-framework-module-http": "^6.2.0",
+    "@equinor/fusion-framework-module-navigation": "^4.0.7",
+    "@equinor/fusion-framework-react": "^7.3.0-next-9f6eec8c5a4cb3dd430fc6bee227442a3dc28bb8",
+    "@equinor/fusion-framework-react-app": "^5.2.10-next-9f6eec8c5a4cb3dd430fc6bee227442a3dc28bb8",
+    "@equinor/fusion-observable": "^8.4.1",
+    "@equinor/fusion-react-styles": "^0.6.2",
+    "@types/react": "^17.0.70",
+    "@types/react-dom": "^17.0.25",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
+    "react-router": "^5.2.0",
+    "react-router-dom": "^5.2.0",
+    "rxjs": "^7.8.1",
+    "styled-components": "^6.1.6",
+    "typesafe-actions": "^5.1.0",
+    "typescript": "^5.1.3",
+    "vite": "^5.2.14",
+    "vite-tsconfig-paths": "^4.2.0"
+  },
+  "devDependencies": {
+    "@equinor/fusion-framework-cli": "^10.0.0-next-499bdf0f6597dbdcfdabd563b94ed08de36bbfe4",
+    "@types/react-router-dom": "^5.3.3",
+    "vite-plugin-static-copy": "^0.17.0"
+  }
+}
