+++ client/apps/fusion-app-loader/src/app-loader/legacy-interopt/static.ts
