+++ client/apps/fusion-app-loader/src/globalResources.ts
