+++ client/apps/fusion-app-loader/tsconfig.json
@@ -1,0 +1,33 @@
+{
+    "compilerOptions": {
+        "rootDir": "src",
+        "baseUrl": "src",
+        "jsx": "react-jsx",
+        "strict": true,
+        "module": "ES2022",
+        "target": "ES6",
+        "incremental": true,
+        "removeComments": true,
+        "preserveConstEnums": true,
+        "inlineSourceMap":true,
+        "moduleResolution": "node",
+        "allowSyntheticDefaultImports": true,
+        "esModuleInterop": true,
+        "stripInternal": true,
+        "skipLibCheck": true,
+        "lib": [
+            "esnext",
+            "dom",
+            "dom.iterable"
+        ],
+        "composite": true,
+        "outDir": "./dist",
+    },
+    "include": [
+        "./src",
+      ],
+      "exclude": [
+        "node_modules",
+        "lib"
+      ]
+}
