+++ client/apps/portal-administration/package.json
@@ -7,7 +7,7 @@
   "main": "src/index.ts",
   "scripts": {
     "build:apps": "fusion-framework-cli app build",
-    "dev": "fusion-framework-cli app dev",
+    "dev:apps": "fusion-framework-cli app dev",
     "docker": "cd .. && sh docker-script.sh app-react",
     "bundle": "fusion-framework-cli app pack"
   },
