--- legacyAppLoader/package.json
@@ -1,45 +1,0 @@
-{
-  "name": "fusion-app-loader",
-  "version": "1.0.0",
-  "description": "",
-  "type": "module",
-  "main": "src/index.ts",
-  "scripts": {
-    "prebuild": "rm -f ../client/packages/portal-client/src/assets/appLegacyLoader.js",
-    "build": "fusion-framework-cli app build && mv ./dist/app-bundle.js ../client/packages/portal-client/src/assets/appLegacyLoader.js",
-    "build:resources": "tsc && vite build",
-    "dev": "fusion-framework-cli app dev -c vite.config.js",
-    "docker": "cd .. && sh docker-script.sh app-react",
-    "update:fusion": "ncu -i -f /fusion/"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@equinor/fusion": "^3.4.13",
-    "@equinor/fusion-components": "^2.12.0",
-    "@equinor/fusion-framework-cli": "^7.0.12",
-    "@equinor/fusion-framework-module-http": "^5.1.1",
-    "@equinor/fusion-framework-module-navigation": "^3.0.6",
-    "@equinor/fusion-framework-react": "^5.3.2",
-    "@equinor/fusion-framework-react-app": "^4.1.9",
-    "@equinor/fusion-observable": "^8.1.2",
-    "@equinor/fusion-react-styles": "^0.5.11",
-    "@types/react": "^17.0.70",
-    "@types/react-dom": "^17.0.25",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "rxjs": "^7.8.1",
-    "styled-components": "^6.1.6",
-    "typesafe-actions": "^5.1.0",
-    "typescript": "^5.1.3",
-    "vite": "^5.2.14",
-    "vite-tsconfig-paths": "^4.2.0"
-  },
-  "devDependencies": {
-    "@types/react-router-dom": "^5.3.3",
-    "vite-plugin-static-copy": "^0.17.0"
-  }
-}
