--- legacyAppLoader/tsconfig.docker.json
@@ -1,24 +1,0 @@
-{
-  "compilerOptions": {
-    "rootDir": "src",
-    "jsx": "react-jsxdev",
-    "strict": true,
-    "module": "ES2022",
-    "target": "ES6",
-    "sourceMap": true,
-    "moduleResolution": "node",
-    "allowSyntheticDefaultImports": true,
-    "esModuleInterop": true,
-    "lib": [
-      "esnext",
-      "dom",
-      "dom.iterable"
-    ],
-  },
-  "include": [
-    "src/**/*"
-  ],
-  "exclude": [
-    "node_modules",
-  ]
-}
