--- legacyAppLoader/tsconfig.json
@@ -1,30 +1,0 @@
-{
-    "compilerOptions": {
-        "rootDir": "src",
-        "baseUrl": "src",
-        "jsx": "react-jsxdev",
-        "strict": true,
-        "module": "ES2022",
-        "target": "ES6",
-        "incremental": true,
-        "removeComments": true,
-        "preserveConstEnums": true,
-
-        "inlineSourceMap":true,
-        "moduleResolution": "node",
-        "allowSyntheticDefaultImports": true,
-        "esModuleInterop": true,
-        "stripInternal": true,
-        "skipLibCheck": true,
-        "lib": [
-            "esnext",
-            "dom",
-            "dom.iterable"
-        ],
-        "composite": true,
-        "outDir": "./dist", 
-    },
-    "include": [
-        "./src",
-      ],
-}
